# contrato/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Pedido - Microserviço",
    "description": "API destinada para uso pela aplicação cliente SOAT - Fast Food",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "tags": [
    {
      "name": "pedido",
      "description": "Endpoints relacionados com pedido"
    }
  ],
  "paths": {
    "/pedido": {
      "post": {
        "tags": [
          "pedido"
        ],
        "summary": "Cria um novo pedido",
        "description": "Permite realizar a criação de um novo pedido",
        "operationId": "criaPedido",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoPedido"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Erro ao criar pedido"
          },
          "401": {
            "description": "Credencial inválida"
          },
          "422": {
            "description": "Parâmetros incompletos"
          }
        }
      }
    },
    "/pedido/{order_id}": {
      "get": {
        "tags": [
          "pedido"
        ],
        "summary": "Obtém detalhes de um pedido pelo ID",
        "description": "Retorna os detalhes de um pedido com base no ID fornecido",
        "operationId": "obterPedidoPorID",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "ID do pedido a ser obtido",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado com sucesso"
          },
          "400": {
            "description": "Erro ao localizar pedido"
          },
          "401": {
            "description": "Credencial inválida"
          },
          "422": {
            "description": "Parâmetros incompletos"
          }
        }
      }
    },
    "/pedidos": {
      "get": {
        "tags": [
          "pedido"
        ],
        "summary": "Obtém todos os pedidos",
        "description": "Retorna todos os pedidos, com suporte a paginação",
        "operationId": "obterTodosOsPedidos",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "Número de registros a serem ignorados no início",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de registros a serem retornados",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedidos encontrados com sucesso"
          },
          "400": {
            "description": "Erro ao localizar pedidos"
          },
          "401": {
            "description": "Credencial inválida"
          },
          "422": {
            "description": "Parâmetros incompletos"
          }
        }
      }
    },
    "/pedido/status/{status}": {
      "get": {
        "tags": [
          "pedido"
        ],
        "summary": "Obtém pedidos pelo status",
        "description": "Retorna pedidos com base no status fornecido",
        "operationId": "obterPedidosPorStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Status do pedido a ser filtrado",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedidos encontrados com sucesso"
          },
          "400": {
            "description": "Status de pedido inválido"
          },
          "401": {
            "description": "Credencial inválida"
          },
          "404": {
            "description": "Nenhum pedido com status informado"
          },
          "422": {
            "description": "Parâmetros incompletos"
          }
        }
      }
    },
    "/pedido_uncompleted": {
      "get": {
        "tags": [
          "pedido"
        ],
        "summary": "Obtém todos os pedidos não concluídos",
        "description": "Retorna todos os pedidos com status que são considerados não concluídos",
        "operationId": "obterPedidosNaoConcluidos",
        "responses": {
          "200": {
            "description": "Pedidos encontrados com sucesso"
          },
          "401": {
            "description": "Credencial inválida"
          },
          "404": {
            "description": "Nenhum pedido pendente"
          },
          "422": {
            "description": "Parâmetros incompletos"
          }
        }
      }
    },
    "/pedido/update_status/{order_id}": {
      "put": {
        "tags": [
          "pedido"
        ],
        "summary": "Atualiza o status do pedido",
        "description": "Atualiza o status do pedido com o novo status fornecido",
        "operationId": "atualizarStatusPedido",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "ID do pedido a ser atualizado",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "new_status",
            "in": "query",
            "description": "Novo status do pedido",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status do pedido atualizado com sucesso"
          },
          "400": {
            "description": "Status de pedido inválido"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    },
    "/pedido/checkout/{order_id}": {
      "put": {
        "tags": [
          "pedido"
        ],
        "summary": "Realiza o checkout do pedido",
        "description": "Atualiza o status de pagamento e de pedido para realizar o checkout",
        "operationId": "realizarCheckoutPedido",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "ID do pedido a ser realizado o checkout",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Checkout do pedido realizado com sucesso"
          },
          "400": {
            "description": "Não é possível realizar o checkout, status de pagamento inválido"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NovoPedido": {
        "type": "object",
        "properties": {
          "cart": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OrderItem"
                }
              }
            }
          },
          "client": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "cpf": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "observation": {
            "type": "string"
          },
          "totalPrice": {
            "type": "number"
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "product": {
            "$ref": "#/components/schemas/OrderProduct"
          },
          "count": {
            "type": "integer"
          },
          "observation": {
            "type": "string"
          }
        }
      },
      "OrderProduct": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "website_auth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "website_auth": []
    }
  ]
}